unessecary login check just added
if this application was a web page this little check would make sence, because it is not possible to hack your way in classified content in Java, but I did it anyway just to be sure.
This is a system that checks when a scene is loaded that the user is logged in. if the user is not logged in he is redirected to the login page.