filling User class with attributes
in order to hanfle the user we will have to instantiate User objects, and this is why I needed at least 4 of the necessary attributes.
Probably whould be correct to have all the attributes at once, but this would make the code overcomplicated and I haven't worked with the comments yet.