Scene7 (Edit User) Scene initiated
the scene is copied from the add user.
Now I will have to adjust the functionality to be able to edit instead of  adding and we will have the necessary result.
I will have to add bit logic though, otherwise the Admin will have edit all the fields on the panel just for simple edit.