oh no...
Almost fone with full CRUD on Users, (without advance safecheck yet), but there is a small detail...
We need to make that the user is not remain admin after the login.
I will try to solve that though the initialization of the login page. One line of code instead of 1 million lines of code. Sounds like a deal.