Now the edited exercises can change recipes as well
This was the most difficult peace of code that I wrote so far. The important thing is that works. and based on this logic we can make the admin to be able to supervise the content of the supervised user.
Extremelly many levels of complexity that can make the application to look at least very impressive.
Anyway... I am freaking done for today.