public recipes. the easy part
To create the placeholder and the isPublic variable on the Recipe class is the easy part.
The most difficult part will be to make everything works without DB and with DB. It is easy to made, but to go through all the users to find the public recipies can produce lage and that is not a good thing.