fixed the issue.
Now it works, hopefully with none or at list very minimum bugs.
The issue was that the DB was giving me null values while  the expected value was int that cant be null. i had to remove the explicit declaration of the recipe ID as foreign key, in order to be able to use the column as I see fit.
Anyway... not it works properly.
Form tomorrow I will start fixing the appereance of the application as well.